:zap: feat: reflact code struct
