+++ src/Layout/demos/single.tsx
@@ -1,0 +1,17 @@
+import { EditorLayout } from '@ant-design/pro-editor';
+
+export default () => {
+  return (
+    <EditorLayout
+      style={{
+        maxWidth: '100%',
+        height: '600px',
+      }}
+      header={false}
+      footer={false}
+      leftPannel={{
+        children: <div>Left Pannel</div>,
+      }}
+    />
+  );
+};
