+++ src/Layout/index.zh-CN.md
@@ -12,4 +12,14 @@ EditorLayout 是我们专门为了编辑器场景制作的编辑器布局组件
 
 ### 基础用法
 
-<code src="./demos/basic.tsx" ></code>
+<code src="./demos/basic.tsx" iframe="600"></code>
+
+### 只需要面板
+
+你可以通过设置 header 和 footer 为 false 来隐藏头和尾的配置
+
+<code src="./demos/single.tsx" ></code>
+
+当然你也可以对某个 pannel 设置 false 来关闭其内容
+
+<code src="./demos/noLeftPannel.tsx" ></code>
